updat

+++ 0x00-hello_world/6-size.c
@@ -1,5 +1,9 @@
 #include <stdio.h>
-/* main -Printing using printf */
+/**
+ * main - Entry point
+ * 
+ * Return: Always 0 (Success)
+ */
 int main(void)
 {
 
@@ -9,11 +13,11 @@ long long int doubleType;
 char charType;
 float floatType;
 
-printf("Size of char: %zu byte\n", sizeof(charType));
-printf("Size of an int: %zu bytes\n", sizeof(intType));
-printf("Size of a long int: %zu bytes\n", sizeof(longIntType));
-printf("Size of a long long int: %zu bytes\n", sizeof(doubleType));
-printf("Size of an float: %zu bytes\n", sizeof(floatType));
+printf("Size of char: %lu byte\n",(unsigned long)sizeof(charType));
+printf("Size of an int: %lu bytes\n", (unsigned long)sizeof(intType));
+printf("Size of a long int: %lu bytes\n", (unsigned long)sizeof(longIntType));
+printf("Size of a long long int: %lu bytes\n", (unsigned long)sizeof(doubleType));
+printf("Size of an float: %lu bytes\n", (unsigned long)sizeof(floatType));
 
 return (0);
 }
